Extracted tabs/tab-action-events to util/add-listener
The function is a good method of generally hooking up to Chrome events.
Changed browser/on-clicked to use it as well.

Renamed browser/on-click -> browser/on-clicked, to match the original
name.

TODO: Decide if the other addListener uses should remain as they are or
if they should use this general function. I'd say we err in the side of
consistency.