BUGFIX: runtime/send-message failed if we didn't send an extensionId
While the extensionId is optional, utils/options->jsparams was filtering
out all nil parameters, which means that if we didn't include an extension
id then the message ended up in place of the extension id and no message
was included.

Defaulting the extensionId to an empty string has the result of the
message being sent to our extension by default.